requirements.txt
label-studio
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
pillow>=9.0.0